+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/repocutter/compare/v0.3.0...HEAD)
 ------------------------------------------------------------------------
+### Added
+- Add `-i/--ignore` argument
 
 [0.3.0](https://github.com/jshwi/repocutter/releases/tag/v0.3.0) - 2023-01-15
 ------------------------------------------------------------------------